refactor: [#661] move E2E tests runner mod
