+++ src/console/ci/e2e/logs_parser.rs
