+++ src/console/ci/e2e/mod.rs
@@ -1,3 +1,4 @@
+//! E2E tests scripts.
 pub mod docker;
 pub mod logs_parser;
 pub mod runner;
