+++ src/console/ci/mod.rs
@@ -1,0 +1,2 @@
+//! Continuos integration scripts.
+pub mod e2e;
