+++ src/console/clients/mod.rs
@@ -1,0 +1,1 @@
+//! Console clients.
