+++ src/console/mod.rs
@@ -1,0 +1,3 @@
+//! Console apps.
+pub mod ci;
+pub mod clients;
