+++ src/lib.rs
@@ -472,8 +472,8 @@
 pub mod app;
 pub mod bootstrap;
 pub mod checker;
+pub mod console;
 pub mod core;
-pub mod e2e;
 pub mod servers;
 pub mod shared;
